[Runtime] Minor update unlock token strategies
1. Use fixed 30 days as unit round. The release amount will cut to 1/2 every 20 rounds.
2. Release amount in a single round is depending on time range. (linear)

+++ src/block.rs
@@ -75,7 +75,7 @@ impl OasisBatchHandler {
 
     pub fn end_batch(&self, ctx: &mut TxnContext) {
         let ectx = runtime_context!(ctx, BlockContext);
-        FundManager::try_unlock(ctx.header.timestamp, &mut ectx.state);
+        FundManager::try_unlock(ctx.header.timestamp as i64, &mut ectx.state);
         info!(ectx.logger, "Commiting state into storage");
         ectx.state.commit().expect("state commit must succeed");
         info!(ectx.logger, "Block finalized");
